eclipse/prova1/src/main/java/prova/prova1/classes/Box.java
package prova.prova1.classes;

import org.springframework.beans.factory.annotation.Value;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@NoArgsConstructor
@AllArgsConstructor
@Data
@Builder
public class Box {

  
    private int x;
    private int y;
    private String name;

    public int getArea(){
        return x*y;
    }

    public int getPerimeter(){
        return x+x+y+y;
    }


    public Box(int x, int y) {
        this.x = x;
        this.y = y;
    }

    public int getX() {
        return this.x;
    }

    public void setX(int x) {
        this.x = x;
    }

    public int getY() {
        return this.y;
    }

    public void setY(int y) {
        this.y = y;
    }

    public String getName() {
        return this.name;
    }

    public void setName(String name) {
        this.name = name;
    }

 
    
}
